+++ src/workflow/Assembler.cpp
@@ -38,7 +38,7 @@ int assembler(int argc, const char **argv, const Command& command) {
     int baseKmerSize = 14;
     par.kmerSize = baseKmerSize;
     par.alphabetSize = Parameters::CLUST_LINEAR_DEFAULT_ALPH_SIZE;
-    for(size_t i = 0; i < par.numIterations; i++){
+    for(int i = 0; i < par.numIterations; i++){
         std::string key = "KMERMATCHER"+SSTR(i)+"_PAR";
     //    par.kmerSize = baseKmerSize - i;
         cmd.addVariable(key.c_str(), par.createParameterString(par.kmermatcher).c_str());
